feat: Add image cropping and background styles
Introduces image cropping functionality using `react-image-crop` to allow users to precisely select the area of their logo to be used in the mockup.

Also expands the available background styles and associated descriptions, providing users with more diverse and descriptive scene options for their mockups. This change refactors the prompt generation in `geminiService` to accommodate the new `BackgroundStyle` type and removes the deprecated `ModelPose` type.